Deployment/requirements.txt
streamlit==1.31.1
Pillow==10.2.0
numpy==1.26.4
ultralytics==8.2.14
matplotlib==3.8.3
pandas==2.2.1
